+++ src/Recyclarr.Cli/Pipelines/QualitySize/PipelinePhases/QualitySizeLogPhase.cs
@@ -28,6 +28,17 @@ public class QualitySizeLogPhase(ILogger log) : ILogPipelinePhase<QualitySizePip
     public void LogPersistenceResults(QualitySizePipelineContext context)
     {
         // Do not check ConfigOutput for null here since that is done for us in the LogConfigPhase method
-        log.Information("Processed Quality Definition: {QualityDefinition}", context.ConfigOutput!.Type);
+        var qualityDefinitionName = context.ConfigOutput!.Type;
+
+        var totalCount = context.TransactionOutput.Count;
+        if (totalCount > 0)
+        {
+            log.Information("Total of {Count} sizes were synced for quality definition {Name}", totalCount,
+                qualityDefinitionName);
+        }
+        else
+        {
+            log.Information("All sizes for quality definition {Name} are already up to date!", qualityDefinitionName);
+        }
     }
 }